requirements.txt
Pillow
numpy
matplotlib
pickle5
scikit-learn
tensorflow
keras